+++ OpenSim/Region/CoreModules/World/Terrain/ITerrainFloodEffect.cs
@@ -32,6 +32,6 @@ namespace OpenSim.Region.CoreModules.World.Terrain
 {
     public interface ITerrainFloodEffect
     {
-        void FloodEffect(ITerrainChannel map, Boolean[,] fillArea, double strength);
+        void FloodEffect(ITerrainChannel map, Boolean[,] fillArea, float strength);
     }
 }
