+++ OpenSim/Region/CoreModules/World/Terrain/ITerrainPaintableEffect.cs
@@ -33,6 +33,6 @@ namespace OpenSim.Region.CoreModules.World.Terrain
 {
     public interface ITerrainPaintableEffect
     {
-        void PaintEffect(ITerrainChannel map, bool[,] allowMask, double x, double y, double z, double strength, double duration, float BrushSize, List<Scene> scene);
+        void PaintEffect(ITerrainChannel map, bool[,] allowMask, float x, float y, float z, float strength, float duration, float BrushSize, List<Scene> scene);
     }
 }
